requirements.txt
python-telegram-bot==20.3
paramiko==2.12.0
psutil==5.9.7
